Reorganize NotImplemented functions on ECS
